scripts/requirements.txt
google-genai
pandas
psycopg2-binary
python-dotenv
beautifulsoup4
backoff
openpyxl
